requirements.txt
Django==5.0.7
pillow==10.4.0
djangorestframework==3.15.2